refactor: move webapp in protocols.http module
This commit moves the web application code in protocols.http module,
close the http client code.
